+++ api/endpoints/management/processes/routes.py
@@ -2,6 +2,7 @@ from flask import jsonify, Blueprint, request
 from flask_jwt_extended import jwt_required
 from werkzeug.exceptions import BadRequest
 from api.core.database import CursorFromPool
+from api.core.query_access import Access
 from api.endpoints.management.processes.models import ProcessModel, DeleteModel
 from api.core.jwt_ext_custom import jwt_required_with_management_claim
 
@@ -144,6 +145,9 @@ def processes_insert():
 @processes_endpoint.route('/api/management/processes/delete', methods=['POST'])
 @jwt_required_with_management_claim()
 def processes_delete():
+    if not Access.to_all_networks():
+        raise BadRequest("Access denied for deleting process")
+
     with CursorFromPool() as cursor:
         model = DeleteModel(**request.json)
         sql = """
